refactor: 서비스 삭제 및 도메인 이동
