+++ src/main/java/domain/CrewFactory.java
@@ -1,13 +1,11 @@
-package service;
+package domain;
 
-import domain.Crew;
-import domain.CrewGroup;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 
-public class CrewService {
-    public CrewGroup createCrewGroup(Map<String, List<LocalDateTime>> crewInitAttendanceDates) {
+public class CrewFactory {
+    public CrewGroup crewGroupOf(Map<String, List<LocalDateTime>> crewInitAttendanceDates) {
         List<String> crewNames = crewInitAttendanceDates.keySet().stream().toList();
         CrewGroup crewGroup = CrewGroup.from(crewNames);
